backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
Pillow>=10.0.0
opencv-python>=4.8.0
pdf2image>=1.16.0
numpy>=1.26.0
paddlepaddle>=2.5.0
paddleocr>=2.7.0
rapidfuzz>=3.0.0
